Allow the recreation of deleted projections (same name)
When deleting a projection, ensure that a $ProjectionDeleted event is
written to $projections-$all
This event allows us to know which projections are actually alive.

When creating a projection, we check for the existence of a
$projections-[projection-name] stream. If this stream exists, we grab the
last known event number, increment it by one and that would be the version
for the projection.